+++ pyramid_oereb/contrib/models/oereblex/forest_distance_lines.py
@@ -80,7 +80,7 @@ class DataIntegration(Base):
         id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         date (datetime.date): The date when this data set was delivered.
-        office_id (int): A foreign key which points to the actual office instance.
+        office_id (str): A foreign key which points to the actual office instance.
         office (pyramid_oereb.standard.models.forest_distance_lines.Office):
             The actual office instance which the id points to.
     """
@@ -88,7 +88,10 @@ class DataIntegration(Base):
     __tablename__ = 'data_integration'
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     date = sa.Column(sa.DateTime, nullable=False)
-    office_id = sa.Column(sa.Integer, sa.ForeignKey(Office.id), nullable=False)
+    office_id = sa.Column(
+        sa.String,
+        sa.ForeignKey(Office.id),
+        nullable=False)
     office = relationship(Office)
     checksum = sa.Column(sa.String, nullable=True)
 
@@ -101,13 +104,14 @@ class ViewService(Base):
         id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         reference_wms (str): The actual url which leads to the desired cartographic representation.
-        legend_at_web (str): A link leading to a wms describing document (png).
+        legend_at_web (dict of str): A multilingual dictionary of links. Keys are the language, values
+            are links leading to a wms describing document (png).
     """
     __table_args__ = {'schema': 'forest_distance_lines'}
     __tablename__ = 'view_service'
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     reference_wms = sa.Column(sa.String, nullable=False)
-    legend_at_web = sa.Column(sa.String, nullable=True)
+    legend_at_web = sa.Column(JSONType, nullable=True)
 
 
 class LegendEntry(Base):
@@ -127,12 +131,12 @@ class LegendEntry(Base):
             legend  entry.
         topic (str): Statement to describe to which public law restriction this legend entry
             belongs.
-        sub_theme (str): Description for sub topics this legend entry might belonging to.
+        sub_theme (dict): Multilingual description for sub topics this legend entry might belonging to.
         other_theme (str): A link to additional topics. It must be like the following patterns
             * ch.{canton}.{topic}  * fl.{topic}  * ch.{bfsnr}.{topic}  This with {canton} as
             the official two letters short version (e.g.'BE') {topic} as the name of the
             topic and {bfsnr} as the municipality id of the federal office of statistics.
-        view_service_id (int): The foreign key to the view service this legend entry is related to.
+        view_service_id (str): The foreign key to the view service this legend entry is related to.
         view_service (pyramid_oereb.standard.models.forest_distance_lines.ViewService):
             The dedicated relation to the view service instance from database.
     """
@@ -144,10 +148,10 @@ class LegendEntry(Base):
     type_code = sa.Column(sa.String(40), nullable=False)
     type_code_list = sa.Column(sa.String, nullable=False)
     topic = sa.Column(sa.String, nullable=False)
-    sub_theme = sa.Column(sa.String, nullable=True)
+    sub_theme = sa.Column(JSONType, nullable=True)
     other_theme = sa.Column(sa.String, nullable=True)
     view_service_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(ViewService.id),
         nullable=False
     )
@@ -163,7 +167,7 @@ class PublicLawRestriction(Base):
             you  don't like it - don't care about.
         information (dict): The multilingual textual representation of the public law restriction.
         topic (str): Category for this public law restriction (name of the topic).
-        sub_theme (str): Textual explanation to subtype the topic attribute.
+        sub_theme (dict): Multilingual textual explanation to subtype the topic attribute.
         other_theme (str): A link to additional topics. It must be like the following patterns
             * ch.{canton}.{topic}  * fl.{topic}  * ch.{bfsnr}.{topic}  This with {canton} as
             the official two letters short version (e.g.'BE') {topic} as the name of the
@@ -176,11 +180,11 @@ class PublicLawRestriction(Base):
         published_from (datetime.date): The date when the document should be available for
             publishing on extracts. This  directly affects the behaviour of extract
             generation.
-        view_service_id (int): The foreign key to the view service this public law restriction is
+        view_service_id (str): The foreign key to the view service this public law restriction is
             related to.
         view_service (pyramid_oereb.standard.models.forest_distance_lines.ViewService):
             The dedicated relation to the view service instance from database.
-        office_id (int): The foreign key to the office which is responsible to this public law
+        office_id (str): The foreign key to the office which is responsible to this public law
             restriction.
         responsible_office (pyramid_oereb.standard.models.forest_distance_lines.Office):
             The dedicated relation to the office instance from database.
@@ -190,7 +194,7 @@ class PublicLawRestriction(Base):
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     information = sa.Column(JSONType, nullable=False)
     topic = sa.Column(sa.String, nullable=False)
-    sub_theme = sa.Column(sa.String, nullable=True)
+    sub_theme = sa.Column(JSONType, nullable=True)
     other_theme = sa.Column(sa.String, nullable=True)
     type_code = sa.Column(sa.String(40), nullable=True)
     type_code_list = sa.Column(sa.String, nullable=True)
@@ -198,7 +202,7 @@ class PublicLawRestriction(Base):
     published_from = sa.Column(sa.Date, nullable=False)
     geolink = sa.Column(sa.Integer, nullable=True)
     view_service_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(ViewService.id),
         nullable=False
     )
@@ -207,7 +211,7 @@ class PublicLawRestriction(Base):
         backref='public_law_restrictions'
     )
     office_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(Office.id),
         nullable=False
     )
@@ -227,12 +231,12 @@ class Geometry(Base):
             generation.
         geo_metadata (str): A link to the metadata which this geometry is based on which delivers
             machine  readable response format (XML).
-        public_law_restriction_id (int): The foreign key to the public law restriction this geometry
+        public_law_restriction_id (str): The foreign key to the public law restriction this geometry
             is  related to.
         public_law_restriction (pyramid_oereb.standard.models.forest_distance_lines
             .PublicLawRestriction): The dedicated relation to the public law restriction instance from
             database.
-        office_id (int): The foreign key to the office which is responsible to this public law
+        office_id (str): The foreign key to the office which is responsible to this public law
             restriction.
         responsible_office (pyramid_oereb.standard.models.forest_distance_lines.Office):
             The dedicated relation to the office instance from database.
@@ -248,7 +252,7 @@ class Geometry(Base):
     geo_metadata = sa.Column(sa.String, nullable=True)
     geom = sa.Column(GeoAlchemyGeometry('LINESTRING', srid=srid), nullable=False)
     public_law_restriction_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(PublicLawRestriction.id),
         nullable=False
     )
@@ -257,7 +261,7 @@ class Geometry(Base):
         backref='geometries'
     )
     office_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(Office.id),
         nullable=False
     )
@@ -272,9 +276,9 @@ class PublicLawRestrictionBase(Base):
     Attributes:
         id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        public_law_restriction_id (int): The foreign key to the public law restriction which bases
+        public_law_restriction_id (str): The foreign key to the public law restriction which bases
             on another  public law restriction.
-        public_law_restriction_base_id (int): The foreign key to the public law restriction which is
+        public_law_restriction_base_id (str): The foreign key to the public law restriction which is
             the  base for the public law restriction.
         plr (pyramid_oereb.standard.models.forest_distance_lines.PublicLawRestriction):
             The dedicated relation to the public law restriction (which bases on) instance from  database.
@@ -285,12 +289,12 @@ class PublicLawRestrictionBase(Base):
     __table_args__ = {'schema': 'forest_distance_lines'}
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     public_law_restriction_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(PublicLawRestriction.id),
         nullable=False
     )
     public_law_restriction_base_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(PublicLawRestriction.id),
         nullable=False
     )
@@ -313,9 +317,9 @@ class PublicLawRestrictionRefinement(Base):
     Attributes:
         id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        public_law_restriction_id (int): The foreign key to the public law restriction which is
+        public_law_restriction_id (str): The foreign key to the public law restriction which is
             refined by  another public law restriction.
-        public_law_restriction_refinement_id (int): The foreign key to the public law restriction
+        public_law_restriction_refinement_id (str): The foreign key to the public law restriction
             which is  the refinement of the public law restriction.
         plr (pyramid_oereb.standard.models.forest_distance_lines.PublicLawRestriction):
             The dedicated relation to the public law restriction (which refines) instance from  database.
@@ -326,12 +330,12 @@ class PublicLawRestrictionRefinement(Base):
     __table_args__ = {'schema': 'forest_distance_lines'}
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     public_law_restriction_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(PublicLawRestriction.id),
         nullable=False
     )
     public_law_restriction_refinement_id = sa.Column(
-        sa.Integer,
+        sa.String,
         sa.ForeignKey(PublicLawRestriction.id),
         nullable=False
     )