+++ pyramid_oereb/lib/records/plr.py
@@ -25,16 +25,7 @@ class EmptyPlrRecord(object):
 class PlrRecord(EmptyPlrRecord):
     """
     Public law restriction record.
-
-    Attributes:
-        part_in_percent (decimal): Part of the property area touched by the restriction in percent.
-        area_share (decimal): Area of the restriction touching the property calculated by the processor.
     """
-
-    # Attributes added or calculated by the processor
-    area_share = None
-
-    part_in_percent = None
 
     def __init__(self, theme, information, law_status, published_from, responsible_office, symbol,
                  view_service, geometries, sub_theme=None, other_theme=None, type_code=None,
@@ -54,7 +45,7 @@ class PlrRecord(EmptyPlrRecord):
                 associated with this record.
             geometries (list of pyramid_oereb.lib.records.geometry.GeometryRecord): List of geometry records
                 associated with this record.
-            sub_theme (unicode): Optional subtopic.
+            sub_theme (dict of unicode or None): Optional subtopic.
             other_theme (unicode): Optional additional topic.
             type_code (unicode): The PLR record's type code (also used by view service).
             type_code_list (unicode): URL to the PLR's list of type codes.
@@ -74,6 +65,9 @@ class PlrRecord(EmptyPlrRecord):
 
         if not isinstance(information, dict):
             warnings.warn('Type of "information" should be "dict"')
+
+        if sub_theme is not None and not isinstance(sub_theme, dict):
+            warnings.warn('Type of "sub_theme" should be "dict"')
 
         assert isinstance(geometries, list)
         assert len(geometries) > 0
@@ -110,6 +104,7 @@ class PlrRecord(EmptyPlrRecord):
         self.area_unit = area_unit
         self.length_unit = length_unit
         self._area_share = None
+        self._part_in_percent = None
         self._length_share = None
         self._nr_of_points = None
         self.symbol = symbol
@@ -159,6 +154,19 @@ class PlrRecord(EmptyPlrRecord):
         return self._area_share
 
     @property
+    def part_in_percent(self):
+        """decimal or None: (decimal): Part of the property area touched by the restriction in percent."""
+        return self._part_in_percent
+
+    @part_in_percent.setter
+    def part_in_percent(self, value):
+        """
+        Args:
+            value (decimal): Part of the property area touched by the restriction in percent.
+        """
+        self._part_in_percent = value
+
+    @property
     def length_share(self):
         """float or None: Returns the summed length of all related geometry records of this PLR."""
         return self._length_share
@@ -194,10 +202,10 @@ class PlrRecord(EmptyPlrRecord):
         area_share = self._sum_area()
         if area_share is None:
             self._area_share = None
-            self.part_in_percent = None
+            self._part_in_percent = None
         else:
             self._area_share = int(round(area_share, 0))
-            self.part_in_percent = round(
+            self._part_in_percent = round(
                 ((float(self._area_share) / float(real_estate.land_registry_area)) * 100),
                 1
             )