+++ pyramid_oereb/lib/records/view_service.py
@@ -25,7 +25,7 @@ class LegendEntryRecord(object):
         type_code_list (unicode): An URL to the type code list.
         theme (pyramid_oereb.lib.records.theme.ThemeRecord): The theme to which the legend entry belongs
             to.
-        sub_theme (unicode): Theme sub category.
+        sub_theme (dict of unicode or None): Theme sub category.
         other_theme (unicode): Additional theme linked to this theme.
         view_service_id (int): The id to the connected view service. This is very important to be able to
             solve bug https://github.com/openoereb/pyramid_oereb/issues/521
@@ -36,6 +36,9 @@ class LegendEntryRecord(object):
 
         if not isinstance(legend_text, dict):
             warnings.warn('Type of "legend_text" should be "dict"')
+
+        if sub_theme is not None and not isinstance(sub_theme, dict):
+            warnings.warn('Type of "sub_theme" should be "dict"')
 
         self.symbol = symbol
         self.legend_text = legend_text
@@ -74,7 +77,8 @@ class ViewServiceRecord(object):
             reference_wms (uri): The link URL to the actual service (WMS)
             layer_index (int): Layer index. Value from -1000 to +1000.
             layer_opacity (float): Opacity of layer. Value from 0.0 to 1.0.
-            legend_at_web (uri): The link URL to the actual legend service (WMS get legend)
+            legend_at_web (dict of str): )A multilingual dictionary of links. Keys are the language, values
+                are links leading to a WMS (get legend) describing document (png).
             legends (list of LegendEntry): A list of all relevant legend entries.
         """
         self.reference_wms = reference_wms