+++ pyramid_oereb/standard/models/groundwater_protection_sites.py
@@ -45,7 +45,7 @@ class Office(Base):
     geometry.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         name (dict): The multilingual name of the office.
         office_at_web (str): A web accessible url to a presentation of this office.
@@ -77,10 +77,10 @@ class DataIntegration(Base):
     able to find out who was the delivering instance.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         date (datetime.date): The date when this data set was delivered.
-        office_id (int): A foreign key which points to the actual office instance.
+        office_id (str): A foreign key which points to the actual office instance.
         office (pyramid_oereb.standard.models.groundwater_protection_sites.Office):
             The actual office instance which the id points to.
     """
@@ -100,12 +100,12 @@ class ReferenceDefinition(Base):
     which are related to an extract but not directly on a special public law restriction situation.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         topic (str): The topic which this definition might be related to.
         canton (str): The canton this definition is related to.
         municipality (int): The municipality this definition is related to.
-        office_id (int): The foreign key constraint which the definition is related to.
+        office_id (str): The foreign key constraint which the definition is related to.
         responsible_office (pyramid_oereb.standard.models.groundwater_protection_sites.Office):
             The dedicated relation to the office instance from database.
     """
@@ -127,7 +127,7 @@ class DocumentBase(Base):
     produce the addressable primary key and to provide the common document attributes.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         text_at_web (dict): A multilingual link which leads to the documents content in the web.
         law_status (str): The status switch if the document is legally approved or not.
@@ -157,7 +157,7 @@ class Document(DocumentBase):
     This represents the main document in the whole system. It is specialized in some sub classes.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         document_type (str): The document type. It must be "LegalProvision", "Law" or "Hint".
         title (dict): The multilingual title or if existing the short title ot his document.
@@ -170,7 +170,7 @@ class Document(DocumentBase):
             the document is  related to the whole canton or even the confederation.
         file (str): The document itself as a binary representation (PDF). It is string but
             BaseCode64 encoded.
-        office_id (int): The foreign key to the office which is in charge for this document.
+        office_id (str): The foreign key to the office which is in charge for this document.
         responsible_office (pyramid_oereb.standard.models.groundwater_protection_sites.Office):
             The dedicated relation to the office instance from database.
     """
@@ -207,11 +207,11 @@ class Article(DocumentBase):
     described as a special part of the whole law document and reflects a dedicated content of this.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         number (str): The number which identifies this article in its parent document.
         text (dict): A simple multilingual string to describe the article or give some related info.
-        document_id (int): The foreign key to the document this article is taken from.
+        document_id (str): The foreign key to the document this article is taken from.
         document_id (pyramid_oereb.standard.models.groundwater_protection_sites.Document):
             The dedicated relation to the document instance from database.
     """
@@ -245,16 +245,17 @@ class ViewService(Base):
     A view service aka WM(T)S which can deliver a cartographic representation via web.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         reference_wms (str): The actual url which leads to the desired cartographic representation.
-        legend_at_web (str): A link leading to a wms describing document (png).
+        legend_at_web (dict of str): A multilingual dictionary of links. Keys are the language, values
+            are links leading to a wms describing document (png).
     """
     __table_args__ = {'schema': 'groundwater_protection_sites'}
     __tablename__ = 'view_service'
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     reference_wms = sa.Column(sa.String, nullable=False)
-    legend_at_web = sa.Column(sa.String, nullable=True)
+    legend_at_web = sa.Column(JSONType, nullable=True)
 
 
 class LegendEntry(Base):
@@ -263,7 +264,7 @@ class LegendEntry(Base):
     :class:`pyramid_oereb.standard.models.groundwater_protection_sites.ViewService`.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         symbol (str): An image with represents the legend entry. This can be png or svg. It is string
             but BaseCode64  encoded.
@@ -274,12 +275,12 @@ class LegendEntry(Base):
             legend  entry.
         topic (str): Statement to describe to which public law restriction this legend entry
             belongs.
-        sub_theme (str): Description for sub topics this legend entry might belonging to.
+        sub_theme (dict): Multilingual description for sub topics this legend entry might belonging to.
         other_theme (str): A link to additional topics. It must be like the following patterns
             * ch.{canton}.{topic}  * fl.{topic}  * ch.{bfsnr}.{topic}  This with {canton} as
             the official two letters short version (e.g.'BE') {topic} as the name of the
             topic and {bfsnr} as the municipality id of the federal office of statistics.
-        view_service_id (int): The foreign key to the view service this legend entry is related to.
+        view_service_id (str): The foreign key to the view service this legend entry is related to.
         view_service (pyramid_oereb.standard.models.groundwater_protection_sites.ViewService):
             The dedicated relation to the view service instance from database.
     """
@@ -291,7 +292,7 @@ class LegendEntry(Base):
     type_code = sa.Column(sa.String(40), nullable=False)
     type_code_list = sa.Column(sa.String, nullable=False)
     topic = sa.Column(sa.String, nullable=False)
-    sub_theme = sa.Column(sa.String, nullable=True)
+    sub_theme = sa.Column(JSONType, nullable=True)
     other_theme = sa.Column(sa.String, nullable=True)
     view_service_id = sa.Column(
         sa.String,
@@ -306,11 +307,11 @@ class PublicLawRestriction(Base):
     The container where you can fill in all your public law restrictions to the topic.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         information (dict): The multilingual textual representation of the public law restriction.
         topic (str): Category for this public law restriction (name of the topic).
-        sub_theme (str): Textual explanation to subtype the topic attribute.
+        sub_theme (dict): Multilingual textual explanation to subtype the topic attribute.
         other_theme (str): A link to additional topics. It must be like the following patterns
             * ch.{canton}.{topic}  * fl.{topic}  * ch.{bfsnr}.{topic}  This with {canton} as
             the official two letters short version (e.g.'BE') {topic} as the name of the
@@ -323,11 +324,11 @@ class PublicLawRestriction(Base):
         published_from (datetime.date): The date when the document should be available for
             publishing on extracts. This  directly affects the behaviour of extract
             generation.
-        view_service_id (int): The foreign key to the view service this public law restriction is
+        view_service_id (str): The foreign key to the view service this public law restriction is
             related to.
         view_service (pyramid_oereb.standard.models.groundwater_protection_sites.ViewService):
             The dedicated relation to the view service instance from database.
-        office_id (int): The foreign key to the office which is responsible to this public law
+        office_id (str): The foreign key to the office which is responsible to this public law
             restriction.
         responsible_office (pyramid_oereb.standard.models.groundwater_protection_sites.Office):
             The dedicated relation to the office instance from database.
@@ -337,7 +338,7 @@ class PublicLawRestriction(Base):
     id = sa.Column(sa.String, primary_key=True, autoincrement=False)
     information = sa.Column(JSONType, nullable=False)
     topic = sa.Column(sa.String, nullable=False)
-    sub_theme = sa.Column(sa.String, nullable=True)
+    sub_theme = sa.Column(JSONType, nullable=True)
     other_theme = sa.Column(sa.String, nullable=True)
     type_code = sa.Column(sa.String(40), nullable=True)
     type_code_list = sa.Column(sa.String, nullable=True)
@@ -365,7 +366,7 @@ class Geometry(Base):
     The dedicated model for all geometries in relation to their public law restriction.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
         law_status (str): The status switch if the document is legally approved or not.
         published_from (datetime.date): The date when the document should be available for
@@ -373,12 +374,12 @@ class Geometry(Base):
             generation.
         geo_metadata (str): A link to the metadata which this geometry is based on which delivers
             machine  readable response format (XML).
-        public_law_restriction_id (int): The foreign key to the public law restriction this geometry
+        public_law_restriction_id (str): The foreign key to the public law restriction this geometry
             is  related to.
         public_law_restriction (pyramid_oereb.standard.models.groundwater_protection_sites
             .PublicLawRestriction): The dedicated relation to the public law restriction instance from
             database.
-        office_id (int): The foreign key to the office which is responsible to this public law
+        office_id (str): The foreign key to the office which is responsible to this public law
             restriction.
         responsible_office (pyramid_oereb.standard.models.groundwater_protection_sites.Office):
             The dedicated relation to the office instance from database.
@@ -416,11 +417,11 @@ class PublicLawRestrictionBase(Base):
     restrictions.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        public_law_restriction_id (int): The foreign key to the public law restriction which bases
+        public_law_restriction_id (str): The foreign key to the public law restriction which bases
             on another  public law restriction.
-        public_law_restriction_base_id (int): The foreign key to the public law restriction which is
+        public_law_restriction_base_id (str): The foreign key to the public law restriction which is
             the  base for the public law restriction.
         plr (pyramid_oereb.standard.models.groundwater_protection_sites.PublicLawRestriction):
             The dedicated relation to the public law restriction (which bases on) instance from  database.
@@ -457,11 +458,11 @@ class PublicLawRestrictionRefinement(Base):
     restrictions.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        public_law_restriction_id (int): The foreign key to the public law restriction which is
+        public_law_restriction_id (str): The foreign key to the public law restriction which is
             refined by  another public law restriction.
-        public_law_restriction_refinement_id (int): The foreign key to the public law restriction
+        public_law_restriction_refinement_id (str): The foreign key to the public law restriction
             which is  the refinement of the public law restriction.
         plr (pyramid_oereb.standard.models.groundwater_protection_sites.PublicLawRestriction):
             The dedicated relation to the public law restriction (which refines) instance from  database.
@@ -497,11 +498,11 @@ class PublicLawRestrictionDocument(Base):
     Meta bucket (join table) for the relationship between public law restrictions and documents.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        public_law_restriction_id (int): The foreign key to the public law restriction which has
+        public_law_restriction_id (str): The foreign key to the public law restriction which has
             relation to  a document.
-        document_id (int): The foreign key to the document which has relation to the public law
+        document_id (str): The foreign key to the document which has relation to the public law
             restriction.
         plr (pyramid_oereb.standard.models.groundwater_protection_sites.PublicLawRestriction):
             The dedicated relation to the public law restriction instance from database.
@@ -536,10 +537,10 @@ class DocumentReference(Base):
     Meta bucket (join table) for the relationship between documents.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        document_id (int): The foreign key to the document which references to another document.
-        reference_document_id (int): The foreign key to the document which is referenced.
+        document_id (str): The foreign key to the document which references to another document.
+        reference_document_id (str): The foreign key to the document which is referenced.
         document (pyramid_oereb.standard.models.groundwater_protection_sites.Document):
             The dedicated relation to the document (which references) instance from database.
         referenced_document (pyramid_oereb.standard.models.groundwater_protection_sites.Document):
@@ -577,11 +578,11 @@ class DocumentReferenceDefinition(Base):
     Meta bucket (join table) for the relationship between documents and the reference definition.
 
     Attributes:
-        id (int): The identifier. This is used in the database only and must not be set manually. If
+        id (str): The identifier. This is used in the database only and must not be set manually. If
             you  don't like it - don't care about.
-        document_id (int): The foreign key to the document which is related to a reference
+        document_id (str): The foreign key to the document which is related to a reference
             definition.
-        reference_definition_id (int): The foreign key to the document which is related to a
+        reference_definition_id (str): The foreign key to the document which is related to a
             reference  definition.
     """
     __tablename__ = 'document_reference_definition'